requirements.txt
gspread==5.7.0
python-dotenv==1.0.0
python-telegram-bot==20.5
fastapi